move legend enums
