--- src/ScottPlot/Legend.cs
@@ -1,29 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Drawing;
-using System.Linq;
-using ScottPlot.Drawing;
-
-namespace ScottPlot
-{
-    public enum legendLocation
-    {
-        none,
-        upperLeft,
-        upperRight,
-        upperCenter,
-        middleLeft,
-        middleRight,
-        lowerLeft,
-        lowerRight,
-        lowerCenter,
-    }
-    public enum shadowDirection
-    {
-        none,
-        upperLeft,
-        upperRight,
-        lowerLeft,
-        lowerRight,
-    }
-}
